% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{abanca2hmbk}
\alias{abanca2hmbk}
\title{Convert abanca to homebank
Convert abanca to homebank}
\usage{
abanca2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
nothing
}
\description{
Convert abanca to homebank
Convert abanca to homebank
}
