% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{abancacredit2hmbk}
\alias{abancacredit2hmbk}
\title{Convert abanca credit to homebank
Convert abanca credit to homebank}
\usage{
abancacredit2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}

\item{outputfile}{Path to the output file}
}
\value{
nothing
}
\description{
Convert abanca credit to homebank
Convert abanca credit to homebank
}
