% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bankinter2hmbk}
\alias{bankinter2hmbk}
\title{Convert bankinter to homebank
Convert bankinter to homebank}
\usage{
bankinter2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
nothing
}
\description{
Convert bankinter to homebank
Convert bankinter to homebank
}
