% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bankinterbusqueda2hmbk}
\alias{bankinterbusqueda2hmbk}
\title{Convert bankinter busqueda to homebank
Convert bankinter, new format (2025-09) to homebank}
\usage{
bankinterbusqueda2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}

\item{outputfile}{Path to the output file}
}
\value{
nothing
}
\description{
Convert bankinter busqueda to homebank
Convert bankinter, new format (2025-09) to homebank
}
