% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bankintercredit2hmbk}
\alias{bankintercredit2hmbk}
\title{Convert bankinter credit to homebank
Convert bankinter credit to homebank}
\usage{
bankintercredit2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
nothing
}
\description{
Convert bankinter credit to homebank
Convert bankinter credit to homebank
}
