% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bbva2hmbk}
\alias{bbva2hmbk}
\title{Convert bbva to homebank
Convert bbva to homebank}
\usage{
bbva2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\description{
Convert bbva to homebank
Convert bbva to homebank
}
