% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bbvacredit2hmbk}
\alias{bbvacredit2hmbk}
\title{Convert bbva credit to homebank
Convert bbva credit to homebank}
\usage{
bbvacredit2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\description{
Convert bbva credit to homebank
Convert bbva credit to homebank
}
