% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bcen2hmbk}
\alias{bcen2hmbk}
\title{Convert b100 to homebank
Convert b100 to homebank}
\usage{
bcen2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}

\item{outputfile}{Path to the output file}
}
\value{
nothing
}
\description{
Convert b100 to homebank
Convert b100 to homebank
}
