% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bmarch2hmbk}
\alias{bmarch2hmbk}
\title{#' Convert banca march to homebank}
\usage{
bmarch2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
#' Convert banca march to homebank
}
