% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{db2hmbk}
\alias{db2hmbk}
\title{Convert db to homebank
Convert db to homebank}
\usage{
db2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\description{
Convert db to homebank
Convert db to homebank
}
