% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{evo2hmbk}
\alias{evo2hmbk}
\title{Convert evobanco to homebank
Convert evobanco to homebank}
\usage{
evo2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}

\item{outputfile}{Path to the output file}
}
\value{
nothing
}
\description{
Convert evobanco to homebank
Convert evobanco to homebank
}
