% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{evocredit2hmbk}
\alias{evocredit2hmbk}
\title{Convert evobanco credit to homebank
Convert evobanco credit to homebank}
\usage{
evocredit2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
nothing
}
\description{
Convert evobanco credit to homebank
Convert evobanco credit to homebank
}
