% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ingbusqueda2hmbk}
\alias{ingbusqueda2hmbk}
\title{Convert ing xls from searches to homebank
Convert ing xls from searches to homebank}
\usage{
ingbusqueda2hmbk(inputfile, outputfile, ingmapcat)
}
\arguments{
\item{inputfile}{Path to the input file}

\item{outputfile}{Path to the output file}

\item{ingmapcat}{path to mapping categories table}
}
\value{
nothing
}
\description{
Convert ing xls from searches to homebank
Convert ing xls from searches to homebank
}
