% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ingnaranja2hmbk}
\alias{ingnaranja2hmbk}
\title{Convert ing cuenta naranja to homebank
Convert ing cuenta naranja to homebank}
\usage{
ingnaranja2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
nothing
}
\description{
Convert ing cuenta naranja to homebank
Convert ing cuenta naranja to homebank
}
