% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ingnomina2hmbk}
\alias{ingnomina2hmbk}
\title{Convert ing cuenta n贸mina to
Convert ing cuenta n贸mina to homebank}
\usage{
ingnomina2hmbk(inputfile, outputfile, ingmapcat)
}
\arguments{
\item{inputfile}{Path to the input file}

\item{ingmapcat}{path to mapping categories table}
}
\value{
nothing
}
\description{
Convert ing cuenta n贸mina to
Convert ing cuenta n贸mina to homebank
}
