% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{multilog2hmbk}
\alias{multilog2hmbk}
\title{Convert multilog to homebank
Convert multilog to homebank}
\usage{
multilog2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
nothing
}
\description{
Convert multilog to homebank
Convert multilog to homebank
}
