% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{myinvestor2hmbk}
\alias{myinvestor2hmbk}
\title{Convert myinvestor to homebank
Convert myinvestor to homebank}
\usage{
myinvestor2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
Convert myinvestor to homebank
Convert myinvestor to homebank
}
