% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{myinvestorcsv2hmbk}
\alias{myinvestorcsv2hmbk}
\title{Convert myinvestor csv to homebank
Convert myinvestor csv to homebank}
\usage{
myinvestorcsv2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
Convert myinvestor csv to homebank
Convert myinvestor csv to homebank
}
