% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{myinvestorxlsx2hmbk}
\alias{myinvestorxlsx2hmbk}
\title{Convert myinvestor xlsx to homebank
Convert myinvestor xlsx to homebank}
\usage{
myinvestorxlsx2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
Convert myinvestor xlsx to homebank
Convert myinvestor xlsx to homebank
}
