% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{n26tohmbk}
\alias{n26tohmbk}
\title{Convert n26 to homebank}
\usage{
n26tohmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
Convert n26 to homebank
}
