% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ob2hmbk}
\alias{ob2hmbk}
\title{#' Convert openbank to homebank}
\usage{
ob2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
#' Convert openbank to homebank
}
