% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{santander2hmbk}
\alias{santander2hmbk}
\title{Convert santander to homebank
Convert santander to homebank}
\usage{
santander2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
Convert santander to homebank
Convert santander to homebank
}
