% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{self2hmbk}
\alias{self2hmbk}
\title{Convert selfbank to homebank
Convert selfbank to homebank}
\usage{
self2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
nothing
}
\description{
Convert selfbank to homebank
Convert selfbank to homebank
}
