% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{wizink2hmbk}
\alias{wizink2hmbk}
\title{Convert wizink to homebank}
\usage{
wizink2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
Convert wizink to homebank
}
