% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{wizinkcredit2hmbk}
\alias{wizinkcredit2hmbk}
\title{#' Convert wizink credit to homebank}
\usage{
wizinkcredit2hmbk(inputfile, outputfile)
}
\arguments{
\item{inputfile}{}

\item{outputfile}{}
}
\value{
nothing
}
\description{
#' Convert wizink credit to homebank
}
